+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Changed
 
 - Eメール送信アクションをメッセージリポジトリに対応
+- 返金アクションをseatInfoSyncCancelに対応
 
 ### Deprecated