+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Added
 
 - 決済サービスのavailableChannel.credentialsにuseSeatInfoSyncCancelを追加
+- 承認にaudienceを拡張
+- 承認オブジェクトをOrganizationRoleに拡張
 
 ### Changed