+++ CHANGELOG.md
@@ -21,6 +21,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - 予約使用アクションのobjectを最適化
 - 注文取引のpotentialActionsを最適化
 - IClientUserを拡張
+- 顧客検索条件
 
 ### Deprecated